% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertICD.R
\name{convertICD}
\alias{convertICD}
\title{Convert ICD code(s)}
\usage{
convertICD(ICDcode, from = "icd10", to = "icd9", flexibility = 1,
  cores = 1)
}
\arguments{
\item{ICDcode}{a vector containing the ICD codes that should be converted. The
code must not contain dots (which are common in some countries representation of the ICD10 system).}

\item{from}{What is the source classification: "icd10" or "icd9"? Defaults to "icd10".}

\item{to}{What is the destination classification: "icd10" or "icd9"? Defaults to "icd9".}

\item{flexibility}{Takes either of the following values: 0, 1, 2:
\itemize{
    \item 0 means no flexibility at all, i.e. the content in ICD9 and ICD10 codes
    must match perfectly; this will probably induce quite many NAs.
    \item 1 means that the function accepts that either the ICD10 or the ICD9
    code more precisely describes the diagnosis.
    \item 2 corresponds to the most flexible conversion, where the function accepts
    that e.g. an ICD10 code is most properly described by several ICD9 codes, or vice versa.
}
Defaults to 1.}

\item{cores}{The number of processor cores to use. Multicore functionality is
available on all unix platforms (e.g. MacOSX, Linux), but not on Windows. Defaults to 1.}
}
\value{
Outputs a character vector containing the converted ICD codes.
}
\description{
Converting ICD codes between the ICD-9 and ICD-10 classifications is not entirely
trivial. There are about twice as many codes in ICD-10 compared to ICD-9. In addition,
several countries have constructed their own variants of the classification systems.
This function will try to convert ICD codes between the two definitions, with a possibility
to specify how flexible the translation should be. If a code is translated and yields multiple
ICD codes in the target definition, only the first code will be returned. 
The function also has multicore support, to reduce computing time when analyzing large datasets. Also, if an
ICD code is not found in the database on its 4-digit representation, the function will
also try to find a 3-digit code by keeping only the first three characters in the code.
It will work only with the 'clean' original versions of ICD, i.e. country-specific codes
have to be converted into a style that the function understands; e.g. "I259" for ICD10, "4149"
for ICD9. This function is in early development, and should be tested more before put to use.
}
\examples{
# Simple example of converting from ICD10 to ICD9:
convertICD("I259")

# Run on several processor cores:
test_codes <- c("I10", "I259", "C820", "E669")
convertICD(test_codes, cores = 4)

}
\author{
Daniel Lindholm
}

